# src/api/swagger.yaml
basePath: /api/v1
...
host: localhost:8086
info:
  contact:
    name: Antelope Team
  description: Antelope bioinformatics pipeline management platform API.
  license:
    name: MIT
  title: Antelope API
  version: "1.0"
paths:
...
  /cab/fastqQuery:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Proxy FASTQ query to CAB
      tags:
      - proxy
  /cab/pipeline:
    post:
      consumes:
      - application/json
      parameters:
      - description: Pipeline submit payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Proxy pipeline submit to CAB
      tags:
      - proxy
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {accessToken}"
    in: header
    name: Authorization
    type: apiKey
  RefreshAuth:
    description: Enter "Bearer {refreshToken}" — used only for /auth/refresh
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"